Making necessary changes on notebook
This commit includes:

1) A reformulation on wiener filter: previously the target wavelet were
generated by the AcousticWave2D operator from Pylops. In this version,
the target wavelet is generated by Devito. This methodology guarantee
that the original and target wavelet were generated in the same way.

2) Plot filtered data: in order to compare the filtering results,
functions to plot filtered data and seismic traces were implemented on
the wiener filter notebook.