Updatings on utils.py and FWI multiscale notebook
A wavelets.py file were created to store the Devito way to generate a
Ricker wavelet and the Fortran way to generate the same wavelet. Bu
studing those implementations, we notice that the Devito wavelet has
really low energy content for low frequencies, which cause some noise by
applying filter wiener for certain frequencies.

To overcome this problem, we implemented the ricker wavelet from
fortran, which have more content energy on low frequencies and not
causing the problem noise anymore. On FWI multiscale jupyter notebook,
you can now choose which one you want to use.

Changes on pylops will be maded to ensure that the fortran wavelet will
be generate as pattern on pylops.