Updating FWI multiscale notebook and adding FWI
This commit consists of a large update on FWI multiscale notebook. The
updates consists of creating a complete framework to run FWI, where the
user just need to set some few parameters and then the FWI will run
according with the setting.

The routine thats program the FWI according with the user is on the
FWI.py file, were uploaded in this commit.